# daveslist-rnqjne/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/message": {
            "get": {
                "tags": [
                    "message"
                ],
                "operationId": "readMessage",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "from message request",
                        "name": "from_user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "tp message request",
                        "name": "to_user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.MessageResp"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.BadRequestErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.UnauthorizedErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/responses.Validate"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.InternalServerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "message"
                ],
                "operationId": "sendMessage",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Login data",
                        "name": "SendMessage",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.MessageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.NoDataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.BadRequestErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.UnauthorizedErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/responses.Validate"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/sell_car/category": {
            "post": {
                "tags": [
                    "category"
                ],
                "operationId": "create-category",
                "parameters": [
                    {
                        "description": "category data data",
                        "name": "category_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.RequestCategory"
                        }
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.NoDataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.BadRequestErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.UnauthorizedErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/sell_car/category/{id}": {
            "delete": {
                "tags": [
                    "category"
                ],
                "operationId": "delete-category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "category id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.NoDataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.BadRequestErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.UnauthorizedErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.InternalServerErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "requests.MessageRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "toUserID": {
                    "type": "integer"
                }
            }
        },
...
        "requests.RequestCategory": {
            "type": "object",
            "required": [
                "category_name",
                "is_public"
            ],
            "properties": {
                "category_name": {
                    "type": "string"
                },
                "is_public": {
                    "type": "boolean"
                }
            }
        },
...
        "responses.BadRequestErrorResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean",
                    "x-order": "0",
                    "example": false
                },
                "code": {
                    "type": "integer",
                    "x-order": "1",
                    "example": 400
                },
                "error": {
                    "type": "object",
                    "properties": {
                        "message": {
                            "type": "string",
                            "x-order": "0",
                            "example": "bad request error message"
                        },
                        "field": {
                            "type": "object",
                            "x-order": "1"
                        }
                    },
                    "x-order": "2"
                }
            }
        },
...
        "responses.InternalServerErrorResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean",
                    "x-order": "0",
                    "example": false
                },
                "code": {
                    "type": "integer",
                    "x-order": "1",
                    "example": 500
                },
                "error": {
                    "type": "object",
                    "properties": {
                        "message": {
                            "type": "string",
                            "x-order": "0",
                            "example": "internal server error"
                        },
                        "field": {
                            "type": "object",
                            "x-order": "1"
                        }
                    },
                    "x-order": "2"
                }
            }
        },
...
        "responses.MessageResp": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "fromUserID": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "toUserID": {
                    "type": "integer"
                }
            }
        },
        "responses.NoData": {
            "type": "object"
        },
        "responses.NoDataResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean",
                    "x-order": "0",
                    "example": true
                },
                "code": {
                    "type": "integer",
                    "x-order": "1"
                },
                "data": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/responses.NoData"
                        }
                    ],
                    "x-order": "2"
                }
            }
        },
...
        "responses.UnauthorizedErrorResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean",
                    "x-order": "0",
                    "example": false
                },
                "code": {
                    "type": "integer",
                    "x-order": "1",
                    "example": 401
                },
                "error": {
                    "type": "object",
                    "properties": {
                        "message": {
                            "type": "string",
                            "x-order": "0",
                            "example": "unauthorized to access the resource"
                        },
                        "field": {
                            "type": "object",
                            "x-order": "1"
                        }
                    },
                    "x-order": "2"
                }
            }
        },
        "responses.Validate": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean",
                    "x-order": "0"
                },
                "code": {
                    "type": "integer",
                    "x-order": "1",
                    "example": 422
                },
                "error": {
                    "type": "object",
                    "properties": {
                        "message": {
                            "type": "string",
                            "x-order": "0",
                            "example": "validate error"
                        },
                        "field": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            },
                            "x-order": "1"
                        }
                    },
                    "x-order": "2"
                }
            }
        }
    }
}